+++ test/test.js
@@ -1,0 +1,25 @@
+/* global
+    describe it beforeEach afterEach
+*/
+import { Optimizer, Optimization, SimpleOptimization } from '../Optimizer.js';
+
+describe('Optimizer', () => {
+
+    let optimizer;
+    beforeEach(() => {
+
+        optimizer = new Optimizer();
+
+    });
+
+
+    it('', ()=>{})
+
+    afterEach(() => {
+
+        optimizer.dispose();
+        optimizer = null;
+
+    });
+
+});